refactor: ask_witness 및 ask_defendant 함수 래퍼 추가 및 구조 변경
