+++ core/chat.py
@@ -1,8 +1,11 @@
 import streamlit as st
 from dotenv import load_dotenv
-from controller import get_ai_response, get_judge_result, get_case_summary, ask_witness, get_witness_profiles, ask_defendant
 import random
 from config import CASE_EXAMPLES
+
+from controller import get_ai_response, get_judge_result, get_case_summary, get_witness_profiles
+from controller import ask_witness_wrapper as ask_witness
+from controller import ask_defendant_wrapper as ask_defendant
 from controller import make_case_judgment_prompt, ask_llm
